Controlling player speed

+++ src/com/quinton/game/entity/projectile/Projectile.java
@@ -8,7 +8,7 @@ import com.quinton.game.graphics.Sprite;
 public abstract class Projectile extends Entity {
 
 	// where projectile spawns
-	final protected int xOrigin, yOrigin;
+	final protected double xOrigin, yOrigin;
 	
 	//overwrites x,y of entity for higher precision
 	protected double x, y;
@@ -23,7 +23,7 @@ public abstract class Projectile extends Entity {
 	
 	protected final Random random = new Random();
 	
-	public Projectile(int x, int y, double dir) {
+	public Projectile(double x, double y, double dir) {
 		xOrigin = x;
 		yOrigin = y;
 		angle = dir;
